move tag out of ast/
