+++ src/types.ts
@@ -1,7 +1,7 @@
 import type Func from './ast/function';
 import type Node from './ast/node';
-import type Tag from './ast/tag';
 import type Variable from './ast/variable';
+import type Tag from './tag';
 
 export type { Node, Tag, Variable };
 export declare type Function = Func;
